Updated timesync and timesync calls
Time of year added to update for timesync. No longer just time of day.

Added a timesync call 1x a day during non-measurement times of year to ensure the unit wakes up when it is supposed to.

Upon power loss or reboot a timesync is attempted regardless of time of year, to ensure the unit starts off with the correct timestamp.